oris-ai-backend/requirements.txt
livekit-agents>=0.12.3
livekit-plugins-deepgram>=0.6.15
livekit-plugins-google>=0.8.1
livekit-plugins-openai[vertex]>=0.10.10
livekit-plugins-silero
livekit-plugins-elevenlabs
livekit-plugins-turn-detector
livekit-plugins-rag>=0.2.3

# Free AI Provider Options
google-generativeai>=0.3.0
edge-tts>=6.1.9
ollama>=0.1.7
anthropic>=0.25.0
livekit-api
python-dotenv~=1.0
requests
aiofile
PyJWT>=2.8.0
llama-index-core
llama-index-embeddings-openai
llama-index-readers-file
torch
transformers